remove amp e2e, integration and unt tests
