+++ cypress/e2e/pages/liveRadio/index.cy.js
@@ -1,13 +1,30 @@
-import runTestsForPage from '../../../support/helpers/runTestsForPage';
-import { testsThatFollowSmokeTestConfig } from './tests';
-import { testsThatFollowSmokeTestConfigForAMPOnly } from './testsForAMPOnly';
-import { testsThatFollowSmokeTestConfigForCanonicalOnly } from './testsForCanonicalOnly';
+import config from '../../../support/config/services';
+import getPaths from '../../../support/helpers/getPaths';
+import serviceHasPageType from '../../../support/helpers/serviceHasPageType';
+import e2eTests from './tests';
+import visitPage from '../../../support/helpers/visitPage';
+import { overrideRendererOnTest } from '../../../support/helpers/onDemandRadioTv';
 
-const testsForPage = {
-  pageType: 'liveRadio',
-  testsThatFollowSmokeTestConfig,
-  testsThatFollowSmokeTestConfigForCanonicalOnly,
-  testsThatFollowSmokeTestConfigForAMPOnly,
-};
-
-runTestsForPage(testsForPage);
+const pageType = 'liveRadio';
+Object.keys(config)
+  .filter(service => serviceHasPageType(service, pageType))
+  .forEach(serviceId => {
+    const { variant, name: service } = config[serviceId];
+    const paths = getPaths(serviceId, pageType);
+    paths.forEach(currentPath => {
+      describe(`${pageType} - ${currentPath}`, () => {
+        beforeEach(() => {
+          Cypress.env('currentPath', currentPath);
+          const newPath = `${Cypress.env(
+            'currentPath',
+          )}${overrideRendererOnTest()}`;
+          visitPage(newPath, pageType);
+        });
+        e2eTests({
+          service,
+          pageType,
+          variant,
+        });
+      });
+    });
+  });
