+++ src/app/pages/LiveRadioPage/index.test.jsx
@@ -4,23 +4,19 @@ import { RequestContextProvider } from '#contexts/RequestContext';
 import * as analyticsUtils from '#lib/analyticsUtils';
 import { ToggleContextProvider } from '#contexts/ToggleContext';
 import { MEDIA_PAGE } from '#app/routes/utils/pageTypes';
-import {
-  render,
-  waitFor,
-} from '../../components/react-testing-library-with-providers';
+import { render } from '../../components/react-testing-library-with-providers';
 import { ServiceContextProvider } from '../../contexts/ServiceContext';
 import LiveRadioPage from './LiveRadioPage';
 import afriquePageData from './fixtureData/afrique';
 import indonesianPageData from './fixtureData/indonesia';
 import gahuzaPageData from './fixtureData/gahuza';
 
-const Page = ({ pageData, service, lang, isAmp = false }) => (
+const Page = ({ pageData, service, lang }) => (
   <BrowserRouter>
     <ToggleContextProvider>
       <ServiceContextProvider service={service} lang={lang}>
         <RequestContextProvider
           bbcOrigin="https://www.test.bbc.com"
-          isAmp={isAmp}
           pageType={MEDIA_PAGE}
           pathname="/pathname"
           service={service}
@@ -43,20 +39,7 @@ jest.mock('../../components/ChartbeatAnalytics', () => {
 describe('Radio Page Main', () => {
   it('should match snapshot for Canonical', () => {
     const { container } = render(
-      <Page
-        pageData={afriquePageData}
-        service="afrique"
-        lang="fr"
-        isAmp={false}
-      />,
-    );
-
-    expect(container).toMatchSnapshot();
-  });
-
-  it('should match snapshot for AMP', () => {
-    const { container } = render(
-      <Page pageData={afriquePageData} service="afrique" lang="fr" isAmp />,
+      <Page pageData={afriquePageData} service="afrique" lang="fr" />,
     );
 
     expect(container).toMatchSnapshot();
@@ -64,12 +47,7 @@ describe('Radio Page Main', () => {
 
   it('should show the title for the Live Radio page', () => {
     const { getByText } = render(
-      <Page
-        pageData={afriquePageData}
-        service="afrique"
-        lang="fr"
-        isAmp={false}
-      />,
+      <Page pageData={afriquePageData} service="afrique" lang="fr" />,
     );
 
     expect(getByText('BBC Afrique Radio')).toBeInTheDocument();
@@ -77,12 +55,7 @@ describe('Radio Page Main', () => {
 
   it('should show the summary for the Live Radio page', () => {
     const { getByText } = render(
-      <Page
-        pageData={afriquePageData}
-        service="afrique"
-        lang="fr"
-        isAmp={false}
-      />,
+      <Page pageData={afriquePageData} service="afrique" lang="fr" />,
     );
 
     expect(getByText('Infos, musique et sports')).toBeInTheDocument();
@@ -90,12 +63,7 @@ describe('Radio Page Main', () => {
 
   it('should show the audio player on canonical', () => {
     const { container } = render(
-      <Page
-        pageData={afriquePageData}
-        service="afrique"
-        lang="fr"
-        isAmp={false}
-      />,
+      <Page pageData={afriquePageData} service="afrique" lang="fr" />,
     );
     const audioPlayerIframeSrc = container
       .querySelector('iframe')
@@ -106,27 +74,9 @@ describe('Radio Page Main', () => {
     );
   });
 
-  it('should show the audio player on AMP', () => {
-    const { container } = render(
-      <Page pageData={afriquePageData} service="afrique" lang="fr" isAmp />,
-    );
-    const audioPlayerIframeSrc = container
-      .querySelector('amp-iframe')
-      .getAttribute('src');
-
-    expect(audioPlayerIframeSrc).toEqual(
-      'https://polling.test.bbc.co.uk/ws/av-embeds/media/bbc_afrique_radio/liveradio/fr/amp?morph_env=live',
-    );
-  });
-
   it('should show the radio schedule for the Live Radio page on canonical', () => {
     const { getByText } = render(
-      <Page
-        pageData={indonesianPageData}
-        service="indonesia"
-        lang="id"
-        isAmp={false}
-      />,
+      <Page pageData={indonesianPageData} service="indonesia" lang="id" />,
     );
     const radioScheduleTitle = getByText('Siaran radio');
     const scheduleWrapper = document.querySelector(
@@ -137,28 +87,9 @@ describe('Radio Page Main', () => {
     expect(radioScheduleTitle).toBeInTheDocument();
   });
 
-  it('should not show the radio schedule for the Live Radio page on AMP', async () => {
-    const { container } = render(
-      <Page
-        pageData={indonesianPageData}
-        service="indonesia"
-        lang="id"
-        isAmp
-      />,
-    );
-
-    const scheduleWrapper = container.querySelector(
-      '[data-e2e="radio-schedule"]',
-    );
-
-    await waitFor(() => {
-      expect(scheduleWrapper).not.toBeInTheDocument();
-    });
-  });
-
   it('should not show the radio schedule for services without a schedule', async () => {
     const { container } = render(
-      <Page pageData={gahuzaPageData} service="gahuza" lang="rw" isAmp />,
+      <Page pageData={gahuzaPageData} service="gahuza" lang="rw" />,
     );
 
     const scheduleWrapper = container.querySelector(
