+++ src/integration/pages/liveRadio/canonicalTests.js
@@ -1,13 +1,32 @@
-import runCrossPlatformTests from './crossPlatformTests';
 import {
   runCoreCanonicalTests,
   runCanonicalAnalyticsTests,
   runRadioScheduleTests,
+  runCommonCrossPlatformTests,
+  runMediaPlayerEmbedLegacyTests,
 } from '../../common';
 
 export default service => {
-  runCrossPlatformTests(service);
   runCoreCanonicalTests();
   runCanonicalAnalyticsTests();
   runRadioScheduleTests({ isAmp: false });
+  runCommonCrossPlatformTests(service);
+  runMediaPlayerEmbedLegacyTests();
+
+  describe('Summary', () => {
+    const summaryEl = document.querySelector('main p');
+    const summaryText = summaryEl.textContent;
+
+    it('should be in the document', () => {
+      expect(summaryEl).toBeInTheDocument();
+    });
+
+    it('should contain text', () => {
+      expect(summaryText).toBeTruthy();
+    });
+
+    it('should match text', () => {
+      expect(summaryText).toMatchSnapshot();
+    });
+  });
 };
