--- src/integration/pages/liveRadio/gahuza/amp.test.js
@@ -1,12 +1,0 @@
-/**
- * @service gahuza
- * @pathname /gahuza/bbc_gahuza_radio/liveradio
- */
-
-import runAmpTests from '../ampTests';
-
-describe('AMP', () => {
-  describe(pageType, () => {
-    runAmpTests(service);
-  });
-});
