--- src/integration/pages/liveRadio/korean/amp.test.js
@@ -1,10 +1,0 @@
-/**
- * @service korean
- * @pathname /korean/bbc_korean_radio/liveradio
- */
-
-import runAmpTests from '../ampTests';
-
-describe('AMP Korean Live Radio Page', () => {
-  runAmpTests(service);
-});
