--- src/integration/pages/liveRadio/sinhala/amp.test.js
@@ -1,10 +1,0 @@
-/**
- * @service sinhala
- * @pathname /sinhala/bbc_sinhala_radio/liveradio
- */
-
-import runAmpTests from '../ampTests';
-
-describe('AMP Sinhala Live Radio Page', () => {
-  runAmpTests(service);
-});
